bot_core/requirements.txt
python-telegram-bot==20.0a4
pyyaml
